apps/ai-service/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python-headless
numpy
requests
python-dotenv
minio
